- Added refresh command on Account Main page

+++ PurchaseManagement/MVVM/Models/Accounts/Statistics.cs
@@ -4,7 +4,7 @@
     {
         public string Day { get; set; }
         public long TotalMoney { get; set; }
-        public long AvgMoney { get; set; }
+        public double AvgMoney { get; set; }
         public int CountMoney { get; set; }
     }
 }
